Address comments & Add func for various sql test
